requirements.txt
opencv-python-headless==4.7.0.68
medpy==0.4.0
tqdm==4.64.1
scipy==1.10.1
simpleitk==2.2.1
numpy==1.23.5
matplotlib==3.7.0
pillow==8.3.2
setuptools==65.6.3
tensorboard==2.12.0
tensorboardX
wheel==0.38.4
pyyaml
scikit-image==0.19.3
pytz
einops
yacs
torch==1.8.0